3d models from 2d images/requirements.txt
Pytorch
pillow
matplotlib
open3d
transformers
spyder